feat(server): MySQL→PostgreSQL 전환 및 마이그레이션 도입 (Phase 0-A)
- TypeORM 드라이버 postgres 전환 (mysql2 제거, pg 추가)
- 엔티티 타입 포팅: FK bigint→int(PK 정합성), tinyint→smallint,
  ContentStatus enum 공유(content_status_enum)
- 쿼리 포팅: LIKE→ILIKE, 계산 alias 정렬→집계식 정렬(PG case-folding 회피)
- TypeORM 마이그레이션 정식 도입: data-source.ts + 초기 스키마 + 조인테이블 created_at 보강
- 로컬 개발 스택 docker-compose.local.yml (PG 5433 / Redis 6380, 호스트 충돌 회피)
- jest src/* 별칭 매핑 추가, 미작동 스텁 spec 제거
- 학습 문서: docs/learning/phase0-mysql-to-postgresql.md

검증: build/test 통과, 공개·인증 엔드포인트 전수 스모크 통과(클라우드 비종속)

+++ server/src/collections/entities/collection.entity.ts
@@ -15,8 +15,8 @@ export class Collection {
   @PrimaryGeneratedColumn('increment')
   id: number;
 
-  // user_id (FK to users.id)
-  @Column({ name: 'user_id', type: 'bigint' })
+  // user_id (FK to users.id). PG FK 타입 정합성을 위해 int 사용.
+  @Column({ name: 'user_id', type: 'int' })
   userId: number;
 
   @Column({ type: 'varchar', length: 30, nullable: false })
